+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Improved Lisp-to-Rust string conversion's performance by making utf-8 validation optional, behind a feature, `utf-8-validation`.
 - Improved Rust-to-Lisp string conversion's performance by not creating a temporary `CString`.
 - Removed `libc` dependency.
+- Deprecated `env.is_not_nil(value)` in favor of `value.is_not_nil()`.
 
 ## [0.9.0] - 2019-07-11
 - `ResultExt` is now a collection of Emacs-specific extension methods for `Result`, instead of a re-export of `failure::ResultExt`.